refactor: rename instrument
